Add files via upload
Task 1 - Number Game

The aim is to create a Java Program in which a user is given K chances to estimate a randomly generated number. The game's rules are as follows: If the guessed number is more than the real number, the program will display the message "guessed number greater than the actual number.